backend/src/pyproject.toml
[tool.poetry]
name = "email-finance-agent"
version = "0.1.0"
description = "AI Email Finance Agent - Backend"
authors = ["AI Agent <ai@finance.agent>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
motor = "^3.4.0"
celery = "^5.3.6"
redis = "^5.0.3"
aioimaplib = "^1.0.1"
msal = "^1.28.0"
httpx = "^0.27.0"
google-api-python-client = "^2.122.0"
google-auth = "^2.29.0"
google-cloud-pubsub = "^2.20.2"
python-magic = "^0.4.27"
PyMuPDF = "^1.24.0"
opencv-python-headless = "^4.9.0.80"
Pillow = "^10.2.0"
numpy = "^1.26.4"
scikit-image = "^0.22.0"
boto3 = "^1.34.69"
aioboto3 = "^12.3.0"
pytesseract = "^0.3.10"
easyocr = "^1.7.1"
transformers = "^4.39.0"
torch = "2.2.1"
huggingface-hub = "^0.22.0"
langchain = "^0.1.13"
langchain-openai = "^0.1.1"
openai = "^1.14.2"
instructor = "^1.2.0"
scikit-learn = "^1.4.1.post1"
nltk = "^3.8.1"
thefuzz = "^0.22.1"
pandas = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.2"
slowapi = "^0.1.8"
sentry-sdk = {extras = ["fastapi"], version = "^1.44.1"}
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.3.3"
black = "^24.3.0"
mypy = "^1.9.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = true
